requirements.txt
flask
flask-sqlalchemy
requests
pillow
pytesseract
reportlab
gunicorn
